requirements.txt
# LangChain + LLM
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0      # kept for OpenRouter compatibility
langchain-google-genai>=2.0.0

# Discord
discord.py>=2.3.0

# Tools
duckduckgo-search>=6.0.0
notion-client>=2.0.0

# Config
pydantic
pydantic-settings
python-dotenv

# Logging
rich

# Testing
pytest
pytest-asyncio
pytest-cov
